webdna-django-server/requirements.txt
amqp==2.3.2
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.0
cffi==1.11.5
Django==2.0.6
django-celery-results==1.0.1
django-cors-headers==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
jsonpickle==0.9.6
kombu==4.2.1
numpy==1.14.5
psycopg2-binary==2.7.5
pycparser==2.18
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
vine==1.1.4
